% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bennu-package.R
\docType{package}
\name{bennu-package}
\alias{bennu-package}
\alias{bennu}
\title{The 'bennu' package.}
\description{
Bayesian Estimation of Naloxone use Number Under-reporting
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sempwn.github.io/bennu/}
  \item Report bugs at \url{https://github.com/sempwn/bennu/issues}
}

}
\author{
\strong{Maintainer}: Mike Irvine \email{mike.irvine@bccdc.ca} (\href{https://orcid.org/0000-0003-4785-8998}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Samantha Bardwell [contributor]
  \item Andrew Johnson [contributor]
}

}
