% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_naloxone.R
\name{est_naloxone}
\alias{est_naloxone}
\title{Run Bayesian estimation of naloxone number under-reporting}
\usage{
est_naloxone(d, psi_vec = c(0.7, 0.2, 0.1), run_estimation = TRUE)
}
\arguments{
\item{d}{data frame with format
\describe{
\item{regions}{unique id for region}
\item{times}{time in months}
\item{Orders}{Kits ordered}
\item{Reported_Used}{Kits reported as used}
\item{Reported_Distributed}{Kits reported as distributed}
\item{region_name}{Optional label for region}
}}

\item{psi_vec}{reporting delay distribution}

\item{run_estimation}{if \code{TRUE} will sample from posterior otherwise will
sample from prior only}
}
\description{
Samples from Bayesian model using input from data frame
}
