% Generated by roxygen2: do not edit by hand
% Please edit documentation in data-raw/create_validation_data.R
\docType{data}
\name{experimental_validation_data}
\alias{experimental_validation_data}
\title{Experimental validation results}
\format{
\subsection{\code{experimental_validation_data}}{

A data frame with 200 rows and 8 columns:
\describe{
\item{.variable}{Model variable}
\item{p50}{Median of the posterior}
\item{p25, p75}{2nd and 3rd quartiles of the posterior}
\item{p05, p95}{1st and 19th ventiles of the posterior}
\item{true_value}{The value used to generate the simulation}
\item{experiment}{Experiment number index}
}
}
}
\usage{
experimental_validation_data
}
\description{
Generated data from validation experiments of simulated data
}
\keyword{datasets}
