% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_model_data.R
\name{generate_model_data}
\alias{generate_model_data}
\title{generate model data for testing purposes}
\usage{
generate_model_data(N_t = 24, region_coeffs = c(5, 0.5), c_region = c(-1, 2))
}
\arguments{
\item{N_t}{number of time-points}

\item{region_coeffs}{vector of coefficients for regions determining kit orders}

\item{c_region}{logit probability of kit use per region}
}
\description{
generate model data for testing purposes
}
