% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-table.R
\name{kit_summary_table}
\alias{kit_summary_table}
\title{Summarize model fit}
\usage{
kit_summary_table(fit, ..., data = NULL, accuracy = 0.01)
}
\arguments{
\item{accuracy}{A number to round to. Use (e.g.) 0.01 to show 2 decimal
places of precision. If NULL, the default, uses a heuristic that should
ensure breaks have the minimum number of digits needed to show the
difference between adjacent values.}
}
\description{
Provides a summary of:
\itemize{
\item Estimated kits distributed
\item Percentage of kits distributed that are reported
\item Estimated kits used
\item percentage of kits used that are reported
\item percentage of kits orders that are used
\item probability kit used if distributed
}
}
