% Generated by roxygen2: do not edit by hand
% Please edit documentation in data-raw/create_missing_data_validation.R
\docType{data}
\name{missing_data_validation}
\alias{missing_data_validation}
\title{Missing data experimental validation results}
\format{
\subsection{\code{missing_data_validation}}{

A data frame with 10 rows and 6 columns:
\describe{
\item{p50}{Median of the posterior}
\item{p25, p75}{2nd and 3rd quartiles of the posterior}
\item{p05, p95}{1st and 19th ventiles of the posterior}
\item{reporting_freq}{The reporting frequency in months}
}
}
}
\usage{
missing_data_validation
}
\description{
Generated data from validation experiments of simulated data
}
\seealso{
Other validation data: 
\code{\link{experimental_validation_data}}
}
\concept{validation data}
\keyword{datasets}
