% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{plot_kit_use}
\alias{plot_kit_use}
\title{Plot of probability of naloxone kit use}
\usage{
plot_kit_use(..., data = NULL)
}
\arguments{
\item{...}{named list of \link{stanfit} objects}

\item{data}{data used for model fitting. Can also include \code{p_use} column
which can be used to plot true values if derived from simulated data.}
}
\value{
\link{ggplot2} object
}
\description{
plot can compare between two different model fits or a single
model fit by region. If data are simulated then can also include in plot.
For more details see the introduction vignette:
\code{vignette("Introduction", package = "bennu")}
}
\seealso{
Other plots: 
\code{\link{kit_summary_table}()}
}
\concept{plots}
