% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bennu-package.R
\docType{package}
\name{bennu-package}
\alias{bennu-package}
\alias{bennu}
\title{The 'bennu' package.}
\description{
Bayesian Estimation of Naloxone use Number Under-reporting
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
