% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_naloxone.R
\name{est_naloxone_vec}
\alias{est_naloxone_vec}
\title{Run Bayesian estimation of naloxone number under-reporting}
\usage{
est_naloxone_vec(
  N_region,
  N_t,
  N_distributed,
  regions,
  times,
  Orders2D,
  Reported_Distributed,
  Reported_Used,
  region_name,
  psi_vec = c(0.7, 0.2, 0.1),
  run_estimation = TRUE,
  rw_type = 1,
  chains = 4,
  iter = 2000,
  seed = 42,
  adapt_delta = 0.85,
  ...
)
}
\arguments{
\item{N_region}{Number of regions}

\item{N_t}{number of time steps}

\item{N_distributed}{Number of samples of reporting for distribution of kits}

\item{regions}{vector (time, region) of regions (coded 1 to N_region)}

\item{times}{vector (time, region) of regions (coded 1 to N_t)}

\item{Orders2D}{vector (time, region) of orders}

\item{Reported_Distributed}{vector (time, region) reported as distributed}

\item{Reported_Used}{vector (time, region) reported as used}

\item{region_name}{bring in region names}

\item{psi_vec}{reporting delay distribution}

\item{run_estimation}{if \code{TRUE} will sample from posterior otherwise will
sample from prior only}

\item{rw_type}{\code{1} - random walk of order one. \code{2} - random walk of order 2.}

\item{chains}{A positive integer specifying the number of Markov chains.
The default is 4.}

\item{iter}{A positive integer specifying the number of iterations
for each chain (including warmup). The default is 2000.}

\item{seed}{Seed for random number generation}

\item{adapt_delta}{(double, between 0 and 1, defaults to 0.8)}

\item{...}{other parameters to pass to \link[rstan:stanmodel-method-sampling]{rstan::sampling}}
}
\description{
Samples from Bayesian model
}
\seealso{
Other inference: 
\code{\link{est_naloxone}()}
}
\concept{inference}
